+++ services/syra-daw-fe/src/ui/atoms/ResizableBox.tsx
@@ -1,8 +1,7 @@
 import { Box, BoxProps } from '@chakra-ui/react';
-import React, { useEffect, useRef } from 'react';
-import { motion, PanInfo, useMotionValue } from 'framer-motion';
-import { DragMode } from '../../types/Ui';
-import useSnapPixelValue from '../../hooks/ui/useSnapPixelValue';
+import React from 'react';
+import { motion } from 'framer-motion';
+import useResizableBox from '../../hooks/ui/useResizableBox';
 
 interface Props extends BoxProps {
   dragHandleWidth?: number;
@@ -28,80 +27,15 @@ const ResizableBox: React.FC<Props> = ({
   onYChanged,
   ...props
 }) => {
-  const width = useMotionValue(baseWidth);
-  const oldWidth = useMotionValue(baseWidth);
-  const x = useMotionValue(baseX);
-  const oldX = useMotionValue(baseX);
-  const boxOffset = useRef(0);
-  const oldBoxOffset = useRef(0);
-  const ref = useRef<HTMLDivElement>(null);
-  const dragMode = useRef(DragMode.MOVE);
-  const snapPixelValue = useSnapPixelValue(0.25);
-
-  useEffect(() => {
-    x.set(baseX);
-    oldX.set(baseX);
-  }, [baseX, x, oldX]);
-
-  useEffect(() => {
-    width.set(baseWidth);
-    oldWidth.set(baseWidth);
-  }, [baseWidth, width, oldWidth]);
-
-  const onMouseDown = (e: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
-    const distance = e.clientX - (ref.current?.getBoundingClientRect().x ?? 0);
-
-    if (distance < dragHandleWidth) {
-      dragMode.current = DragMode.START_HANDLE;
-    } else if (distance > oldWidth.get() - dragHandleWidth) {
-      dragMode.current = DragMode.END_HANDLE;
-    } else {
-      dragMode.current = DragMode.MOVE;
-    }
-  };
-
-  const onDrag = (_: MouseEvent, { offset }: PanInfo) => {
-    if (lockDrag) {
-      onYChanged && onYChanged(offset.y);
-
-      x.set(oldX.get());
-
-      return;
-    }
-
-    const snappedOffset = snapPixelValue(offset.x);
-
-    switch (dragMode.current) {
-      case DragMode.END_HANDLE:
-        width.set(oldWidth.get() + snappedOffset);
-        x.set(oldX.get());
-        break;
-      case DragMode.START_HANDLE:
-        console.log('oldBoxOffset', oldBoxOffset.current);
-        console.log('snappedOffset', snappedOffset);
-
-        if (oldBoxOffset.current + snappedOffset >= 0 || allowOverExtendingStart) {
-          width.set(oldWidth.get() - snappedOffset);
-          x.set(oldX.get() + snappedOffset);
-          boxOffset.current = oldBoxOffset.current + snappedOffset;
-        }
-        break;
-      case DragMode.MOVE:
-        x.set(oldX.get() + snappedOffset);
-        break;
-    }
-  };
-
-  const onDragEnd = () => {
-    if (x.get() < 0) {
-      x.set(0);
-    }
-
-    onPositionChanged(x.get(), width.get(), boxOffset.current);
-    oldWidth.set(width.get());
-    oldX.set(x.get());
-    oldBoxOffset.current = boxOffset.current;
-  };
+  const { onDragEnd, onMouseDown, onDrag, width, x, ref } = useResizableBox(
+    baseWidth,
+    baseX,
+    dragHandleWidth,
+    onPositionChanged,
+    onYChanged,
+    lockDrag,
+    allowOverExtendingStart,
+  );
 
   return (
     <motion.div
